Reverted utils renaming (not necessary as they are internal modules)
